build db script added
