+++ config/connection.js
@@ -2,26 +2,7 @@
 
 require('dotenv').config();
 const Sequelize = require('sequelize');
-const mysql = require('mysql2');
-
-async function createDB() {
-    return new Promise ( (res, rej)  => {
-        const connection = mysql.createConnection({
-            host: 'localhost',
-            user: process.env.DB_USER,
-            password: process.env.DB_PASSWORD,
-        });
-        connection.query( 
-            'CREATE DATABASE IF NOT EXISTS nerdquiz_db',
-            function(err, results) {
-                if(err) {console.log(err)}
-                else console.log(results)
-            });
-        connection.end();
-    })
-}
-
-createDB();
+// const mysql = require('mysql2');
 
 // // connects to db and created db if not exists
 // const connection = mysql.createConnection({
