+++ seeds/index.js
@@ -7,16 +7,6 @@ const seedUserScores = require('./userScore-seeds');
 
 const sequelize = require('../config/connection');
 
-// const createnerdDB = async () => {
-//     await sequelize.query(
-//         'CREATE DATABASE IF NOT EXISTS nerdquiz_db;',
-//         function(err, results) {
-//             if(err) {console.log(err)}
-//             else console.log(results)
-//         }
-//     )
-// };
-
 const seedAll = async () => {
     await sequelize.sync({ force: true });
     console.log('\n**//---- DB synced ----\\\\**\n');
@@ -35,5 +25,4 @@ const seedAll = async () => {
     process.exit(0)
 };
 
-// createnerdDB();
 seedAll();
